requirements.txt
faiss-cpu==1.11.0
google-genai==1.15.0
google-generativeai==0.8.5
langchain==0.3.25
langchain-community==0.3.24
langchain-google-genai==2.0.10
python-dotenv==1.1.0
streamlit==1.45.1